Implement VPN connection
Moved away from trying to automatically enter the saved credentials,
because I couldn't get it to work.
Instead, this adds the option to use `pass` to get the saved credentials
which are simply printed before connecting, so one can just copy/paste
them from within one terminal.

This also changes the initialization to let the user choose between
saving the credentials in plaintext, or the `pass` command to use to
decrypt the credentials.